Knowledge Checks/Week 4/Korean Recipes & Top YT Videos/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
google-api-python-client
python-dotenv
requests
requests-mock
bs4
